backend/requirements.txt
fastapi
uvicorn
google-cloud-logging
pydantic
python-dotenv
protobuf